Plano AEE - não renderizar o plano novamente quando trocar de turma
